move key rotation command to ipfs key rotate
