+++ cmd/ipfs/ipfs.go
@@ -22,7 +22,6 @@ var commandsClientCmd = commands.CommandsCmd(Root)
 var localCommands = map[string]*cmds.Command{
 	"daemon":   daemonCmd,
 	"init":     initCmd,
-	"rotate":   rotateCmd,
 	"commands": commandsClientCmd,
 }
